dxFilterBuilder: Fix after review
